+++ test/general/sycl_iterator_buf2.pass.cpp
@@ -13,48 +13,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "support/test_config.h"
+#include "sycl_iterator.pass.h"
 
-#include _PSTL_TEST_HEADER(execution)
-#include _PSTL_TEST_HEADER(algorithm)
-#include _PSTL_TEST_HEADER(numeric)
-#include _PSTL_TEST_HEADER(memory)
-#include _PSTL_TEST_HEADER(iterator)
-
-#include "support/utils.h"
-#include "oneapi/dpl/pstl/utils.h"
-
-#include <cmath>
-#include <type_traits>
-
-using namespace TestUtils;
-
-//This macro is required for the tests to work correctly in CI with tbb-backend.
 #if TEST_DPCPP_BACKEND_PRESENT
-#include "support/utils_sycl.h"
-
-struct Flip
-{
-    std::int32_t val;
-    Flip(std::int32_t y) : val(y) {}
-    template <typename T>
-    T
-    operator()(const T& x) const
-    {
-        return val - x;
-    }
-};
-struct Plus
-{
-    template <typename T, typename U>
-    T
-    operator()(const T x, const U y) const
-    {
-        return x + y;
-    }
-};
-
-using namespace oneapi::dpl::execution;
 
 DEFINE_TEST(test_uninitialized_copy)
 {
@@ -75,9 +36,8 @@ DEFINE_TEST(test_uninitialized_copy)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from uninitialized_copy");
     }
@@ -102,9 +62,8 @@ DEFINE_TEST(test_uninitialized_copy_n)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_copy_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from uninitialized_copy_n");
     }
@@ -128,9 +87,8 @@ DEFINE_TEST(test_uninitialized_move)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_move(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from uninitialized_move");
     }
@@ -155,9 +113,8 @@ DEFINE_TEST(test_uninitialized_move_n)
         update_data(host_keys, host_vals);
 
         ::std::uninitialized_move_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
                     "wrong effect from uninitialized_move_n");
@@ -179,10 +136,8 @@ DEFINE_TEST(test_uninitialized_fill)
 
         ::std::uninitialized_fill(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + (n / 3), first1 + (n / 2),
                                   value);
+        wait_and_throw(exec);
 
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2),
                                  value),
@@ -204,9 +159,8 @@ DEFINE_TEST(test_uninitialized_fill_n)
         auto value = T1(2);
 
         ::std::uninitialized_fill_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, value + 1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, value + 1),
                     "wrong effect from uninitialized_fill_n");
@@ -232,9 +186,8 @@ DEFINE_TEST(test_uninitialized_default_construct)
 
         ::std::uninitialized_default_construct(make_new_policy<new_kernel_name<Policy, 0>>(exec),
                                              first1 + (n / 3), first1 + (n / 2));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), exp_value),
                     "wrong effect from uninitialized_default_construct");
@@ -259,9 +212,8 @@ DEFINE_TEST(test_uninitialized_default_construct_n)
         host_keys.update_data();
 
         ::std::uninitialized_default_construct_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, exp_value),
                     "wrong effect from uninitialized_default_construct_n");
@@ -285,9 +237,8 @@ DEFINE_TEST(test_uninitialized_value_construct)
 
         ::std::uninitialized_value_construct(make_new_policy<new_kernel_name<Policy, 0>>(exec),
                                            first1 + (n / 3), first1 + (n / 2));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get() + (n / 3), host_keys.get() + (n / 2), T1{}),
                     "wrong effect from uninitialized_value_construct");
@@ -311,9 +262,8 @@ DEFINE_TEST(test_uninitialized_value_construct_n)
         host_keys.update_data();
 
         ::std::uninitialized_value_construct_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         EXPECT_TRUE(check_values(host_keys.get(), host_keys.get() + n, T1{}),
                     "wrong effect from uninitialized_value_construct_n");
@@ -339,9 +289,7 @@ DEFINE_TEST(test_transform_unary)
         update_data(host_keys, host_vals);
 
         ::std::transform(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + n / 2, last1, first2 + n / 2, Flip(7));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
 
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n / 2, value + 1),
@@ -369,9 +317,8 @@ DEFINE_TEST(test_transform_binary)
         host_keys.update_data();
 
         ::std::transform(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first1, first2, Plus());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, T1(6)),
                     "wrong effect from transform_binary");
@@ -395,9 +342,8 @@ DEFINE_TEST(test_replace_copy)
         host_keys.update_data();
 
         ::std::replace_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, value, T1(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value + 1),
                     "wrong effect from replace_copy");
@@ -422,9 +368,8 @@ DEFINE_TEST(test_replace_copy_if)
 
         ::std::replace_copy_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2,
                              oneapi::dpl::__internal::__equal_value<T1>(value), T1(value + 1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value + 1),
                     "wrong effect from replace_copy_if");
@@ -449,9 +394,8 @@ DEFINE_TEST(test_copy)
         update_data(host_keys, host_vals);
 
         ::std::copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
                     "wrong effect from copy");
@@ -477,9 +421,8 @@ DEFINE_TEST(test_copy_n)
         update_data(host_keys, host_vals);
 
         ::std::copy_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, n, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value), "wrong effect from copy_n");
     }
@@ -504,9 +447,8 @@ DEFINE_TEST(test_move)
         update_data(host_keys, host_vals);
 
         ::std::move(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         EXPECT_TRUE(check_values(host_vals.get(), host_vals.get() + n, value),
                     "wrong effect from move");
@@ -540,9 +482,7 @@ DEFINE_TEST(test_adjacent_difference)
 
         // test with custom functor
         ::std::adjacent_difference(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, __f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
 
         {
             retrieve_data(host_keys, host_vals);
@@ -562,9 +502,7 @@ DEFINE_TEST(test_adjacent_difference)
         host_vals.update_data();
 
         ::std::adjacent_difference(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
 
         retrieve_data(host_keys, host_vals);
 
@@ -597,9 +535,8 @@ DEFINE_TEST(test_transform_reduce_binary)
 
         auto result =
             ::std::transform_reduce(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first1, T1(42));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == n + 42, "wrong effect from transform_reduce (2 binary)");
     }
 };
@@ -626,14 +563,13 @@ DEFINE_TEST(test_adjacent_find)
         // check with no adjacent equal elements
         Iterator result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         Iterator expected = last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #1 no elements)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                     << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check with the last adjacent element
         ::std::size_t max_dis = n;
@@ -644,14 +580,13 @@ DEFINE_TEST(test_adjacent_find)
         }
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         expected = max_dis > 1 ? last - 2 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #2 the last element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check with an adjacent element
         max_dis = n;
@@ -664,24 +599,24 @@ DEFINE_TEST(test_adjacent_find)
         }
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         expected = max_dis > 1 ? it - 1 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #3 middle element)");
-#    if _ONEDPL_DEBUG_SYCL
+
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
+
         // check with an adjacent element (no predicate)
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #4 middle element (no predicate))");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check with the first adjacent element
         max_dis = n;
@@ -692,14 +627,13 @@ DEFINE_TEST(test_adjacent_find)
         }
         result = ::std::adjacent_find(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         expected = max_dis > 1 ? first : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from adjacent_find (Test #5 the first element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
     }
 };
 
@@ -725,14 +659,13 @@ DEFINE_TEST(test_is_sorted_until)
         // check sorted
         Iterator result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, comp);
         Iterator expected = last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected, "wrong effect from is_sorted_until (Test #1 sorted sequence)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check unsorted: the last element
         ::std::size_t max_dis = n;
@@ -743,15 +676,14 @@ DEFINE_TEST(test_is_sorted_until)
         }
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, comp);
         expected = max_dis > 1 ? last - 1 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected,
                     "wrong effect from is_sorted_until (Test #2 unsorted sequence - the last element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check unsorted: the middle element
         max_dis = n;
@@ -764,27 +696,26 @@ DEFINE_TEST(test_is_sorted_until)
         }
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last, comp);
         expected = it;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected,
                     "wrong effect from is_sorted_until (Test #3 unsorted sequence - the middle element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
+
         // check unsorted: the middle element (no predicate)
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(
             result == expected,
             "wrong effect from is_sorted_until (Test #4 unsorted sequence - the middle element (no predicate))");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         // check unsorted: the first element
         if (n > 1)
@@ -794,15 +725,14 @@ DEFINE_TEST(test_is_sorted_until)
         }
         result = ::std::is_sorted_until(make_new_policy<new_kernel_name<Policy, 4>>(exec), first, last, comp);
         expected = n > 1 ? first + 1 : last;
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result == expected,
                     "wrong effect from is_sorted_until (Test #5 unsorted sequence - the first element)");
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         ::std::cout << "got: [" << ::std::distance(first, result) << "], "
                   << "expected: [" << ::std::distance(first, expected) << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
     }
 };
 
@@ -831,15 +761,13 @@ DEFINE_TEST(test_equal)
         auto expected  = new_end - new_start > 0;
         auto result = ::std::equal(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1 + new_start,
                                    first1 + new_end, first2 + new_start);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(expected == result, "wrong effect from equal with 3 iterators");
         result = ::std::equal(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1 + new_start, first1 + new_end,
                               first2 + new_start, first2 + new_end);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(expected == result, "wrong effect from equal with 4 iterators");
     }
 };
@@ -864,43 +792,37 @@ DEFINE_TEST(test_find_if)
         {
             auto res0 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1,
                                        [n](T1 x) { return x == n - 1; });
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from find_if_0");
             res0 = ::std::find(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, first1, T1(1));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from find_0");
         }
         // find_if
         auto res1 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1,
                                    [n](T1 x) { return x == n - 1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE((res1 - first1) == n - 1, "wrong effect from find_if_1");
 
         auto res2 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1,
                                    [](T1 x) { return x == -1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res2 == last1, "wrong effect from find_if_2");
 
         auto res3 = ::std::find_if(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1,
                                    [](T1 x) { return x % 2 == 0; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res3 == first1, "wrong effect from find_if_3");
 
         //find
         auto res4 = ::std::find(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, T1(-1));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res4 == last1, "wrong effect from find");
     }
 };
@@ -929,18 +851,16 @@ DEFINE_TEST(test_find_first_of)
 
             auto res =
                 ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1, "Wrong effect from find_first_of_1");
         }
         else if (n >= 2 && n < 10)
         {
             auto res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1,
                                             first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_first_of_2");
 
             // No matches
@@ -948,9 +868,8 @@ DEFINE_TEST(test_find_first_of)
             host_vals.update_data();
 
             res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_first_of_3");
         }
         else if (n >= 10)
@@ -966,9 +885,8 @@ DEFINE_TEST(test_find_first_of)
             host_keys.update_data();
 
             auto res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1 + pos2, "Wrong effect from find_first_of_4");
 
             // Add second match
@@ -976,9 +894,8 @@ DEFINE_TEST(test_find_first_of)
             host_keys.update_data();
 
             res = ::std::find_first_of(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1 + pos1, "Wrong effect from find_first_of_5");
         }
     }
@@ -1005,14 +922,12 @@ DEFINE_TEST(test_search)
         if (n == 1)
         {
             auto res0 = ::std::search(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from search_00");
             res0 = ::std::search(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0 == first1, "wrong effect from search_01");
         }
         auto res1 = ::std::search(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, last2);
@@ -1021,16 +936,14 @@ DEFINE_TEST(test_search)
         {
             // first n-10 elements of the subsequence are at the beginning of first sequence
             auto res2 = ::std::search(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2 + 10, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res2 - first1 == 5, "wrong effect from search_2");
         }
         // subsequence consists of one element (last one)
         auto res3 = ::std::search(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, last1 - 1, last1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(last1 - res3 == 1, "wrong effect from search_3");
 
         // first sequence contains 2 almost similar parts
@@ -1040,9 +953,8 @@ DEFINE_TEST(test_search)
             host_keys.update_data();
 
             auto res4 = ::std::search(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, first2 + 5, first2 + 6);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res4 == first1, "wrong effect from search_4");
         }
     }
@@ -1070,9 +982,8 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, n - start, T(11));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start, "wrong effect from search_1");
         }
         // Search for sequence in the middle
@@ -1084,17 +995,15 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 1>>(exec), first, last, end - start, T(22));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start, "wrong effect from search_20");
 
             // Search for sequence of lesser size
             res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last,
                                 ::std::max(end - start - 1, (size_t)1), T(22));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start, "wrong effect from search_21");
         }
         // Search for sequence at the beginning
@@ -1105,9 +1014,8 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last, end, T(33));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first, "wrong effect from search_3");
         }
         // Search for sequence that covers the whole range
@@ -1116,33 +1024,29 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 4>>(exec), first, last, n, T(44));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first, "wrong effect from search_4");
         }
         // Search for sequence which is not there
         {
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 5>>(exec), first, last, 2, T(55));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last, "wrong effect from search_50");
 
             // Sequence is there but of lesser size(see search_n_3)
             res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 6>>(exec), first, last, (n / 3 + 1), T(33));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last, "wrong effect from search_51");
         }
 
         // empty sequence case
         {
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 7>>(exec), first, first, 1, T(5 + n - 1));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first, "wrong effect from search_6");
         }
         // 2 distinct sequences, must find the first one
@@ -1160,9 +1064,8 @@ DEFINE_TEST(test_search_n)
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 8>>(exec), first, last,
                                      ::std::min(end1 - start1, end2 - start2), T(66));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res - first == start1, "wrong effect from search_7");
         }
 
@@ -1175,9 +1078,8 @@ DEFINE_TEST(test_search_n)
             host_keys.update_data();
 
             auto res = ::std::search_n(make_new_policy<new_kernel_name<Policy, 9>>(exec), first + 1, last, seq_len, T(77));
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last, "wrong effect from search_8");
         }
     }
@@ -1204,14 +1106,12 @@ DEFINE_TEST(test_mismatch)
         if (n == 1)
         {
             auto res0 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, first1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0.first == first1 && res0.second == first2, "wrong effect from mismatch_00");
             res0 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res0.first == first1 && res0.second == first2, "wrong effect from mismatch_01");
         }
         auto res1 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
@@ -1220,9 +1120,8 @@ DEFINE_TEST(test_mismatch)
         {
             // first n-10 elements of the subsequence are at the beginning of first sequence
             auto res2 = ::std::mismatch(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2 + 5, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res2.first == last1 - 5 && res2.second == last2, "wrong effect from mismatch_2");
         }
     }
@@ -1248,9 +1147,8 @@ DEFINE_TEST(test_transform_inclusive_scan)
         auto res1 = ::std::transform_inclusive_scan(
             make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, ::std::plus<T1>(),
             [](T1 x) { return x * 2; }, value);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last2, "wrong result from transform_inclusive_scan_1");
 
         retrieve_data(host_keys, host_vals);
@@ -1307,9 +1205,8 @@ DEFINE_TEST(test_transform_exclusive_scan)
         auto res1 =
             ::std::transform_exclusive_scan(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2,
                                           T1{}, ::std::plus<T1>(), [](T1 x) { return x * 2; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last2, "wrong result from transform_exclusive_scan");
 
         auto ii = T1(0);
@@ -1345,9 +1242,8 @@ DEFINE_TEST(test_copy_if)
 
         auto res1 = ::std::copy_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2,
                                    [](T1 x) { return x > -1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last2, "wrong result from copy_if_1");
 
         host_vals.retrieve_data();
@@ -1364,9 +1260,8 @@ DEFINE_TEST(test_copy_if)
 
         auto res2 = ::std::copy_if(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2,
                                  [](T1 x) { return x % 2 == 1; });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res2 == first2 + (last2 - first2) / 2, "wrong result from copy_if_2");
 
         host_vals.retrieve_data();
@@ -1399,9 +1294,8 @@ DEFINE_TEST(test_remove)
 
         auto pos = (last - first) / 2;
         auto res1 = ::std::remove(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, T1(222 + pos));
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last - 1, "wrong result from remove");
 
         host_keys.retrieve_data();
@@ -1436,9 +1330,8 @@ DEFINE_TEST(test_remove_if)
         auto pos = (last - first) / 2;
         auto res1 = ::std::remove_if(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last,
                                    [=](T1 x) { return x == T1(222 + pos); });
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(res1 == last - 1, "wrong result from remove_if");
 
         host_keys.retrieve_data();
@@ -1481,20 +1374,19 @@ DEFINE_TEST(test_unique_copy)
         auto result_first = first2;
         auto result_last =
             ::std::unique_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, result_first, f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         auto result_size = result_last - result_first;
 
         std::int64_t expected_size = (n - 1) / 4 + 1;
 
         // check
         bool is_correct = result_size == expected_size;
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         if (!is_correct)
             ::std::cout << "buffer size: got " << result_last - result_first << ", expected " << expected_size
                       << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         host_vals.retrieve_data();
         auto host_first2 = host_vals.get();
@@ -1531,19 +1423,18 @@ DEFINE_TEST(test_unique)
         // invoke
         auto f = [](IteratorValueType a, IteratorValueType b) { return a == b; };
         auto result_last = ::std::unique(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, f);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         auto result_size = result_last - first;
 
         std::int64_t expected_size = (n - 1) / 4 + 1;
 
         // check
         bool is_correct = result_size == expected_size;
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         if (!is_correct)
             ::std::cout << "buffer size: got " << result_last - first << ", expected " << expected_size << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
 
         host_keys.retrieve_data();
         auto host_first1 = host_keys.get();
@@ -1552,10 +1443,10 @@ DEFINE_TEST(test_unique)
             if (*(host_first1 + i) != i + 1)
             {
                 is_correct = false;
-#    if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
                 ::std::cout << "got: " << *(host_first1 + i) << "[" << i << "], "
                           << "expected: " << i + 1 << "[" << i << "]" << ::std::endl;
-#    endif
+#endif // _ONEDPL_DEBUG_SYCL
             }
             EXPECT_TRUE(is_correct, "wrong effect from unique");
         }
@@ -1579,9 +1470,8 @@ DEFINE_TEST(test_is_heap_until)
         host_keys.update_data();
 
         auto actual = ::std::is_heap_until(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         // first element is always a heap
         EXPECT_TRUE(actual == first + 1, "wrong result of is_heap_until_1");
 
@@ -1593,9 +1483,8 @@ DEFINE_TEST(test_is_heap_until)
         host_keys.update_data();
 
         actual = ::std::is_heap_until(make_new_policy<new_kernel_name<Policy, 2>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(actual == (first + n / 2), "wrong result of is_heap_until_2");
 
         host_keys.retrieve_data();
@@ -1603,9 +1492,8 @@ DEFINE_TEST(test_is_heap_until)
         host_keys.update_data();
 
         actual = ::std::is_heap_until(make_new_policy<new_kernel_name<Policy, 3>>(exec), first, last);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(actual == last, "wrong result of is_heap_until_3");
     }
 };
@@ -1627,14 +1515,13 @@ DEFINE_TEST(test_stable_sort)
         host_keys.update_data();
 
         ::std::stable_sort(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         {
             host_keys.retrieve_data();
             auto host_first1 = host_keys.get();
 
-#           if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
             for (int i = 0; i < n; ++i)
             {
                 if (host_first1[i] != value + i)
@@ -1643,17 +1530,17 @@ DEFINE_TEST(test_stable_sort)
                               << ::std::endl;
                 }
             }
-#           endif
+#endif // _ONEDPL_DEBUG_SYCL
+
             EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n), "wrong effect from stable_sort_1");
         }
 
         ::std::stable_sort(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, ::std::greater<T1>());
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_keys.retrieve_data();
         auto host_first1 = host_keys.get();
-#       if _ONEDPL_DEBUG_SYCL
+#if _ONEDPL_DEBUG_SYCL
         for (int i = 0; i < n; ++i)
         {
             if (host_first1[i] != value + n - 1 - i)
@@ -1662,7 +1549,8 @@ DEFINE_TEST(test_stable_sort)
                             << ", got = " << host_first1[i] << ::std::endl;
             }
         }
-#       endif
+#endif // _ONEDPL_DEBUG_SYCL
+
         EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n, ::std::greater<T1>()),
                     "wrong effect from stable_sort_3");
     }
@@ -1693,9 +1581,7 @@ DEFINE_TEST(test_partial_sort)
         {
             auto end1 = first1 + end_idx;
             ::std::partial_sort(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, end1, last1);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
 
             // Make sure that elements up to end are sorted and remaining elements are bigger
             // than the last sorted one.
@@ -1712,9 +1598,8 @@ DEFINE_TEST(test_partial_sort)
         if (end_idx > last1 - first1)
         {
             ::std::partial_sort(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, last1);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             host_keys.retrieve_data();
             auto host_first1 = host_keys.get();
             EXPECT_TRUE(::std::is_sorted(host_first1, host_first1 + n), "wrong effect from partial_sort_2");
@@ -1750,9 +1635,8 @@ DEFINE_TEST(test_partial_sort_copy)
 
             auto last_sorted =
                 ::std::partial_sort_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, end2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             auto host_first1 = host_keys.get();
             auto host_first2 = host_vals.get();
 
@@ -1771,9 +1655,7 @@ DEFINE_TEST(test_partial_sort_copy)
         {
             auto last_sorted =
                 ::std::partial_sort_copy(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
 
             auto host_first1 = host_keys.get();
             auto host_first2 = host_vals.get();
@@ -1815,9 +1697,8 @@ DEFINE_TEST(test_find_end)
 
             // Empty subsequence
             auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, first2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_end_1");
 
             return;
@@ -1832,9 +1713,8 @@ DEFINE_TEST(test_find_end)
 
             // No subsequence
             auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, first2 + n / 2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == last1, "Wrong effect from find_end_2");
 
             // Whole sequence is matched
@@ -1842,9 +1722,8 @@ DEFINE_TEST(test_find_end)
             host_keys.update_data();
 
             res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-            exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+            wait_and_throw(exec);
+
             EXPECT_TRUE(res == first1, "Wrong effect from find_end_3");
 
             return;
@@ -1864,9 +1743,8 @@ DEFINE_TEST(test_find_end)
 
                 auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2,
                                            first2 + n / 5);
-#if _PSTL_SYCL_TEST_USM
-                exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+                wait_and_throw(exec);
+
                 EXPECT_TRUE(res == first1, "Wrong effect from find_end_4");
             }
 
@@ -1880,9 +1758,8 @@ DEFINE_TEST(test_find_end)
 
                 auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, first2,
                                            first2 + n / 5);
-#if _PSTL_SYCL_TEST_USM
-                exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+                wait_and_throw(exec);
+
                 EXPECT_TRUE(res == first1 + 2 * n / 5, "Wrong effect from find_end_5");
             }
 
@@ -1895,9 +1772,8 @@ DEFINE_TEST(test_find_end)
 
                 auto res = ::std::find_end(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, first2,
                                            first2 + n / 5);
-#if _PSTL_SYCL_TEST_USM
-                exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+                wait_and_throw(exec);
+
                 EXPECT_TRUE(res == first1 + 4 * n / 5, "Wrong effect from find_end_6");
             }
         }
@@ -1934,9 +1810,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 1.1: S1 == S2 && len(S1) == len(S2)
         bool is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1,
                                                           last1, first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0, "wrong effect from lex_compare Test 1.1: S1 == S2 && len(S1) == len(S2)");
@@ -1944,9 +1819,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 1.2: S1 == S2 && len(S1) < len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1 - 1,
                                                    first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != 1)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 1" << ::std::endl;
         EXPECT_TRUE(is_less_res == 1, "wrong effect from lex_compare Test 1.2: S1 == S2 && len(S1) < len(S2)");
@@ -1954,9 +1828,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 1.3: S1 == S2 && len(S1) > len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 2>>(exec), first1, last1, first2,
                                                    last2 - 1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0, "wrong effect from lex_compare Test 1.3: S1 == S2 && len(S1) > len(S2)");
@@ -1970,9 +1843,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 2.1: S1 < S2 (PRE-LAST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 3>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         bool is_less_exp = n > 1 ? 1 : 0;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -1982,9 +1854,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 2.2: S1 < S2 (PRE-LAST ELEMENT) && len(S1) > len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 4>>(exec), first1, last1, first2,
                                                    last2 - 1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
         EXPECT_TRUE(is_less_res == is_less_exp,
@@ -1999,9 +1870,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 3.1: S1 > S2 (PRE-LAST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 5>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0,
@@ -2010,9 +1880,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 3.2: S1 > S2 (PRE-LAST ELEMENT) && len(S1) < len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 6>>(exec), first1, last1 - 1,
                                                    first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         is_less_exp = n > 1 ? 0 : 1;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -2026,9 +1895,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 4.1: S1 < S2 (FIRST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 7>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != 1)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 1" << ::std::endl;
         EXPECT_TRUE(is_less_res == 1, "wrong effect from lex_compare Test 4.1: S1 < S2 (FIRST) && len(S1) == len(S2)");
@@ -2036,9 +1904,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 4.2: S1 < S2 (FIRST ELEMENT) && len(S1) > len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 8>>(exec), first1, last1, first2,
                                                    last2 - 1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         is_less_exp = n > 1 ? 1 : 0;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -2052,9 +1919,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 5.1: S1 > S2 (FIRST ELEMENT) && len(S1) == len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 9>>(exec), first1, last1, first2,
                                                    last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         if (is_less_res != 0)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected 0" << ::std::endl;
         EXPECT_TRUE(is_less_res == 0, "wrong effect from lex_compare Test 5.1: S1 > S2 (FIRST) && len(S1) == len(S2)");
@@ -2062,9 +1928,8 @@ DEFINE_TEST(test_lexicographical_compare)
         // CHECK 5.2: S1 > S2 (FIRST ELEMENT) && len(S1) < len(S2)
         is_less_res = ::std::lexicographical_compare(make_new_policy<new_kernel_name<Policy, 10>>(exec), first1, last1 - 1,
                                                    first2, last2, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         is_less_exp = n > 1 ? 0 : 1;
         if (is_less_res != is_less_exp)
             ::std::cout << "N=" << n << ": got " << is_less_res << ", expected " << is_less_exp << ::std::endl;
@@ -2093,9 +1958,8 @@ DEFINE_TEST(test_swap_ranges)
 
         Iterator2 actual_return = ::std::swap_ranges(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2);
 
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         bool check_return = (actual_return == last2);
         EXPECT_TRUE(check_return, "wrong result of swap_ranges");
         if (check_return)
@@ -2141,9 +2005,7 @@ DEFINE_TEST(test_nth_element)
         // invoke
         auto comp = ::std::less<T1>{};
         ::std::nth_element(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, middle1, last1, comp);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
 
         retrieve_data(host_keys, host_vals);
 
@@ -2188,9 +2050,8 @@ DEFINE_TEST(test_reverse_copy)
         ::std::reverse(local_copy.begin(), local_copy.end());
 
         ::std::reverse_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, last, result_first);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         auto host_first2 = host_vals.get();
         for (int i = 0; i < n; ++i)
@@ -2218,9 +2079,8 @@ DEFINE_TEST(test_rotate_copy)
         ::std::rotate(local_copy.begin(), local_copy.begin() + 1, local_copy.end());
 
         ::std::rotate_copy(make_new_policy<new_kernel_name<Policy, 0>>(exec), first, first + 1, last, result_first);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         host_vals.retrieve_data();
         for (int i = 0; i < n; ++i)
             EXPECT_TRUE(local_copy[i] == host_vals.get()[i], "wrong effect from rotate_copy");
@@ -2264,9 +2124,8 @@ DEFINE_TEST(test_includes)
         host_vals.update_data(nb);
 
         auto result = ::std::includes(make_new_policy<new_kernel_name<Policy, 0>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
+
         EXPECT_TRUE(result, "wrong effect from includes a, b");
 
         host_vals.retrieve_data();
@@ -2274,9 +2133,7 @@ DEFINE_TEST(test_includes)
         host_vals.update_data(nc);
 
         result = ::std::includes(make_new_policy<new_kernel_name<Policy, 1>>(exec), first1, last1, first2, last2);
-#if _PSTL_SYCL_TEST_USM
-        exec.queue().wait_and_throw();
-#endif // _PSTL_SYCL_TEST_USM
+        wait_and_throw(exec);
 
         EXPECT_TRUE(!result, "wrong effect from includes a, c");
     }
